move logic to webapps.py
